ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
lightgbm==4.3.0
joblib==1.4.2
httpx==0.27.0
python-dotenv==1.0.1
